services/cobot-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
dobot_py
